flexalloc: Move zns code into znd files
Also removed some printf statements and added asserts where we fail to allocate
memory and if we miscalculate the number of zones being managed.
